add structure of project
