+++ src/config.js
@@ -1,0 +1,5 @@
+import { Telegraf } from "telegraf";
+const BOT_TOKEN = "5687253547:AAHqiWGBpGg_dgW_kLgamFBEJkio-eqceI8",
+  bot = new Telegraf(BOT_TOKEN);
+
+export default bot;
