+++ src/info.js
@@ -1,0 +1,88 @@
+import bot from "./config.js";
+
+bot.action("infoSymbol", (ctx) => {
+  ctx.replyWithHTML(
+    "Користувач вводить символ за яким він бажає отримати детальну інформацію. Символ виглядає як пара двох символів, альткоїна(наприклад BTC) та стейблкоїна(USDT). Стейблкоїн USDT постійно дорівнюї 1$, це для того щоб легче та краще розуміти ціну визначеної криптовалюти. Після введення символа, корисувач отримує інформації, у вигляді:\nСимвол;\nЦіна;\nНапрямок ціни;\nНайвища ціна за 24 год;\nНайменша ціна за 24 год;\nроцентна зміна ціни відносно 1 год;\nЦіна 24 год тому;\nОборот за 24 год;\nОб'єм за 24 год.\n\n"
+  );
+});
+bot.action("infoOrderBook", (ctx) => {
+  ctx.replyWithHTML(
+    "Order Book(книга замовлень) - це список замовлень з ордерів що створюють користувачі за визначеної ціної та об'ємом, які виражаються у $ США. У боті користувач може запросити книгу замовлень за визначеним символом та отримати по 25 ордерів на купівлю та продаж. "
+  );
+});
+bot.action("infoQueryKline", (ctx) => {
+  ctx.replyWithHTML(
+    "Query Kline(клин запит) - це функція яка виводить ціну за символом, яка була визначена якимось проміжком часу раніше. Користувач робить запит за параметрами які йому потрібні, кожен параметр обов'язковий. Перший з них це символ для пошуку. Другий це interval(інтервал часу), тобто, це який проміжок часу між якими буде визначатись ціни для виведення, наприклад, інтервал у 60 секунд, це значить, що різница у часі між першої ціною і другої буде у 60 секунд. Третій параметр, це відколи буде починатися пошук цін, будто місяць, тиждень, день, година або навіть 5 секунд. Та останній параметр це кількість запитів який отримає користувач, є обмеження у 25 запитів."
+  );
+});
+bot.action("infoLatestBigDeal", (ctx) => {
+  ctx.replyWithHTML(
+    "Latest Big Deal(остання велика згода) - це функція яка виводить згоди, сума яких була перевищена за 500.000$. Ця функція потребує введення символу з обов'язковим стейблкоїном USDT. Також, вона буде корисна тим, що досить великі згоди впливають на ринок і буде дуже цікаво на яку ціну була згода що змінила потік ринку."
+  );
+});
+bot.action("infoGetWalletBalance", (ctx) => {
+  ctx.replyWithHTML(
+    "Get Wallet Balance(отримання балансу гаманця) - ця функція виводить у подробицях баланс гаманця користувача за ринком деривіативів. Ця функція буде дуже зручно тим користувачам, у яких досить слабкий зв'язок Інтернету, для того щоб зайти на сайт/додаток біржі ByBit для перегляду балансу свого гаманця"
+  );
+});
+bot.action("infoGetActiveOrder", (ctx) => {
+  ctx.replyWithHTML(
+    "Get Active Order(отримання активного замовлення) - ця функція виконує виведення усіх активних замовлень користувача на ринку диривативів, вона потребує обов'язкого введення API Key та API Secret юзера, без них даною функцієї не можна буде користуватись"
+  );
+});
+bot.action("infoPlaceActiveOrder", (ctx) => {
+  ctx.replyWithHTML("");
+});
+bot.action("infoCancelActiveOrder", (ctx) => {
+  ctx.replyWithHTML(
+    "Cancel Active Order(відміна активного замовлення) - ця функція яка відміняє активне замовлення на ринку диривативів за назвою символа та ідентифікаційним номером самого ордера, який можна отримати з функції GetActiveOrder, але дана функція може відміняти ті ордери, які не були виконані або виконані частично."
+  );
+});
+bot.action("infoCancelAllActiveOrder", (ctx) => {
+  ctx.replyWithHTML(
+    "Cancel All Active Orders(відміна всіх активних замовлень) - назва функції сама каже за себе, то що вона відміняє усі активні ордери користувача на ринку диривативів, які не були виконані або виконані частично"
+  );
+});
+
+export function getInfoText() {
+  return bot.command("info", (ctx) => {
+    ctx.replyWithHTML(
+      "Натисніть на кнопку потрібної Вам функції, для отримання більш детальної інформації:",
+      {
+        reply_markup: {
+          inline_keyboard: [
+            [{ text: "Symbol", callback_data: "infoSymbol" }],
+            [{ text: "Order Book", callback_data: "infoOrderBook" }],
+            [{ text: "Query Kline", callback_data: "infoQueryKline" }],
+            [{ text: "Latest Big Deal", callback_data: "infoLatestBigDeal" }],
+            [
+              {
+                text: "Get Wallet Balance",
+                callback_data: "infoGetWalletBalance",
+              },
+            ],
+            [{ text: "Get Active Order", callback_data: "infoGetActiveOrder" }],
+            [
+              {
+                text: "Place Active Order",
+                callback_data: "infoPlaceActiveOrder",
+              },
+            ],
+            [
+              {
+                text: "Cancel Active Order",
+                callback_data: "infoCancelActiveOrder",
+              },
+            ],
+            [
+              {
+                text: "Cancel All Active Order",
+                callback_data: "infoCancelAllActiveOrder",
+              },
+            ],
+          ],
+        },
+      }
+    );
+  });
+}
