+++ src/inverse/inverse.js
@@ -1,0 +1,3 @@
+// import bot from "../config.js";
+import choose, {chooseButton} from "./commands/choose.js"
+
